Fix review bugs and code quality issues
- Change reviewDuration from Integer to Long in CardUpdateRequest,
  ReviewLog entity, and schema (integer → bigint) to avoid overflow
- Remove silent error swallowing in fetchSessionStats — let errors
  propagate instead of returning null for network/auth failures
- Replace CSS class selectors (.stat-value) in E2E tests with
  semantic getByLabel('value') queries
- Extract duplicated stats grid template into reusable ng-template
- Convert cardShownAt from mutable field to signal

+++ server/src/main/java/io/github/mucsi96/learnlanguage/entity/ReviewLog.java
@@ -63,5 +63,5 @@ public class ReviewLog {
     @Column(nullable = false)
     private LocalDateTime review;
 
-    private Integer reviewDuration;
+    private Long reviewDuration;
 }
