+++ server/src/main/java/io/github/mucsi96/learnlanguage/model/CardUpdateRequest.java
@@ -26,6 +26,6 @@ public class CardUpdateRequest {
     private LocalDateTime lastReview;
     private Integer learningPartnerId;
     private Integer rating;
-    private Integer reviewDuration;
+    private Long reviewDuration;
     private Boolean flagged;
 }
